# spec/openapi.yaml
openapi: 3.0.1
info:
  title: YumYum Delivery Microservice
  description: Delivery Microservice for YumYum
  version: 1.0.0
servers:
  - url: 'http://localhost:8081/v1'
...
paths:
...
  /private-courier/{courierId}:
    get:
      tags:
        - Courier
      summary: Get vendor for courier
      description: Get vendor-courier relation for the vendor that hired the provided courier
      operationId: getCourier
      parameters:
        - name: userId
          in: header
          required: true
          schema:
            $ref: '#/components/schemas/id'
          example: 1015
        - name: courierId
          in: path
          required: true
          schema:
            $ref: '#/components/schemas/id'
          example: 1816
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PrivateCourier'
              example:
                courierId: 1816
                vendorId: 158
        '400':
          description: Bad Request
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
              example:
                code: 400
                message: Bad Request
        '403':
          description: Forbidden
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
              example:
                code: 403
                message: Forbidden
        '404':
          description: Not Found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
              example:
                code: 404
                message: Not Found
    put:
      tags:
        - Courier
      summary: Update vendor for courier
      description: This sets the vendor that hired the courier
      operationId: updateCourier
      parameters:
        - name: userId
          in: header
          required: true
          schema:
            $ref: '#/components/schemas/id'
          example: 1015
        - name: courierId
          in: path
          required: true
          schema:
            $ref: '#/components/schemas/id'
          example: 1816
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                vendorId:
                  $ref: '#/components/schemas/id'
              example:
                vendorId: 158
              required:
                - vendorId
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PrivateCourier'
              example:
                courierId: 1816
                vendorId: 158
        '400':
          description: Bad Request
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
              example:
                code: 400
                message: Bad Request
        '403':
          description: Forbidden
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
              example:
                code: 403
                message: Forbidden
        '404':
          description: Not Found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
              example:
                code: 404
                message: Not Found
  /courier/{courierId}/getLocation:
    get:
      tags:
        - Courier
      summary: Get courier location
      description: Get real time location of courier. This will be fetched from a GPS service
      operationId: getCourierLocation
      parameters:
        - name: userId
          in: header
          required: true
          schema:
            $ref: '#/components/schemas/id'
          example: 1015
        - name: courierId
          in: path
          required: true
          schema:
            $ref: '#/components/schemas/id'
          example: 1816
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Location'
        '400':
          description: Bad Request
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
              example:
                code: 400
                message: Bad Request
        '403':
          description: Forbidden
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
              example:
                code: 403
                message: Forbidden
        '404':
          description: Not Found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
              example:
                code: 404
                message: Not Found
  /admin/default-radius:
    get:
      tags:
        - Admin
      summary: Get default delivery radius
      description: Get the app-wide default delivery radius. This is used if the vendor does not have a custom delivery radius
      operationId: getCurrentDefaultRadius
      parameters:
        - name: userId
          in: header
          required: true
          schema:
            $ref: '#/components/schemas/id'
          example: 0001
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                type: object
                properties:
                  radius:
                    type: integer
                    example: 50
        '400':
          description: Bad Request
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
              example:
                code: 400
                message: Bad Request
        '403':
          description: Forbidden
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
              example:
                code: 403
                message: Forbidden
        '404':
          description: Not Found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
              example:
                code: 404
                message: Not Found
    put:
      tags:
        - Admin
      summary: Set default radius
      description: Updates the default radius of all vendors without a specified radius
      operationId: setDefaultRadius
      parameters:
        - name: userId
          in: header
          required: true
          schema:
            $ref: '#/components/schemas/id'
          example: 0001
        - name: radius
          in: query
          required: true
          schema:
            type: object
            properties:
              radius:
                type: integer
                example: 55
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                type: object
                properties:
                  radius:
                    type: integer
                    example: 55
        '400':
          description: Bad Request
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
              example:
                code: 400
                message: Bad Request
        '403':
          description: Forbidden
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
              example:
                code: 403
                message: Forbidden
        '404':
          description: Not Found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
              example:
                code: 404
                message: Not Found
  /admin/analytics:
    get:
      tags:
        - Admin
      summary: Get analytics
      description: Fetches the analytics for the admins
      operationId: getAnalytics
      parameters:
        - name: userId
          in: header
          required: true
          schema:
            $ref: '#/components/schemas/id'
          example: 0001
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Analytics'
              example:
                avgDeliveryTime: 25
                driverEfficiency: 98.05
                completedDeliveries: 54
                issues: [ late, late, food arrived cold ]

        '400':
          description: Bad Request
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
              example:
                code: 400
                message: Bad Request
        '403':
          description: Forbidden
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
              example:
                code: 403
                message: Forbidden
        '404':
          description: Not Found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
              example:
                code: 404
                message: Not Found
components:
  schemas:
    id:
      type: integer
      description: Schema for our ids, keeps them all constant
      format: int64
    PrivateCourier:
      type: object
      x-class-extra-annotation: "@javax.persistence.Entity  @javax.persistence.Table(name = \"couriers\")"
      description: Courier that is privately hired by a vendor. Can only be assigned to the vendor that hired him
      properties:
        courierId:
          type: integer
          format: int64
          x-field-extra-annotation: "@javax.persistence.Id @javax.persistence.GeneratedValue(strategy = javax.persistence.GenerationType.IDENTITY)"
        vendorId:
          $ref: '#/components/schemas/id'
...
    Error:
      type: object
      x-class-extra-annotation: "@javax.persistence.Embeddable"
      x-field-extra-annotation: "@javax.persistence.Embedded"
      description: Error with error code and description
      properties:
        code:
          type: integer
        message:
          type: string
      example:
        code: 400
        message: Bad Request
    Analytics:
      type: object
      description: Analytics object that stores relevant information for our admins
      properties:
        avgDeliveryTime:
          type: integer
        driverEfficiency:
          type: integer
        completedDeliveries:
          type: integer
        issues:
          type: array
          items:
            type: string
...
    Location:
      type: object
      x-class-extra-annotation: "@javax.persistence.Embeddable"
      x-field-extra-annotation: "@javax.persistence.Embedded"
      description: Address
      properties:
        country:
          type: string
        city:
          type: string
        address:
          type: string
        postalCode:
          type: string
      example:
        country: Netherlands
        city: Delft
        address: Mekelweg 4
        postalCode: 2628 CD
...